Calling r.raise_for_status soley, removed exception handling from get() functionality.  Moved Unauthorized exception handling tests from unit to integration to appease Github CI.
